wb-homa-adc.conf.d folder processing is added
